qcacld-2.0: Configure FW not to wakeup APPS for RRM action frame
If the DUT connects to 11k enabled AP and enter WoW mode, then
the AP is keep sending RRM beacon reports (Action frames) which are
waking up the APPS and leading to battery drain.

The fix is to configure the firmware to not to wake up APPS
for RRM action frame.

Change-Id: I2368be4f641b8b8684a458daea52a81fa671200b
CRs-Fixed: 1115560
Bug: 33507597
Signed-off-by: Srinivas Girigowda <sgirigow@codeaurora.org>
Signed-off-by: ahmedradaideh <ahmed.radaideh@gmail.com>